requirements.txt
torch==1.9.0+cu111
torchvision==0.10.0+cu111
torchaudio==0.9.0
dgl-cu111
pysam
scikit-learn
concurrent_log_handler
regex
pandas
